Bump java quarkus plugin scaffold to v1beta
