+++ pkg/quarkus/v1beta/scaffolds/api.go
@@ -20,9 +20,9 @@ import (
 	"sigs.k8s.io/kubebuilder/v3/pkg/model/resource"
 	"sigs.k8s.io/kubebuilder/v3/pkg/plugins"
 
-	"github.com/operator-framework/java-operator-plugins/pkg/quarkus/v1alpha/scaffolds/internal/templates/controller"
-	"github.com/operator-framework/java-operator-plugins/pkg/quarkus/v1alpha/scaffolds/internal/templates/model"
-	"github.com/operator-framework/java-operator-plugins/pkg/quarkus/v1alpha/util"
+	"github.com/operator-framework/java-operator-plugins/pkg/quarkus/v1beta/scaffolds/internal/templates/controller"
+	"github.com/operator-framework/java-operator-plugins/pkg/quarkus/v1beta/scaffolds/internal/templates/model"
+	"github.com/operator-framework/java-operator-plugins/pkg/quarkus/v1beta/util"
 )
 
 type apiScaffolder struct {
