+++ pkg/quarkus/v1beta/scaffolds/init.go
@@ -18,11 +18,11 @@ import (
 	"os"
 	"path/filepath"
 
-	"github.com/operator-framework/java-operator-plugins/pkg/quarkus/v1alpha/util"
+	"github.com/operator-framework/java-operator-plugins/pkg/quarkus/v1beta/util"
 	"sigs.k8s.io/kubebuilder/v3/pkg/config"
 	"sigs.k8s.io/kubebuilder/v3/pkg/machinery"
 
-	"github.com/operator-framework/java-operator-plugins/pkg/quarkus/v1alpha/scaffolds/internal/templates"
+	"github.com/operator-framework/java-operator-plugins/pkg/quarkus/v1beta/scaffolds/internal/templates"
 	"sigs.k8s.io/kubebuilder/v3/pkg/plugins"
 )
