+++ pkg/quarkus/v1beta/scaffolds/internal/templates/controller/controller.go
@@ -19,7 +19,7 @@ import (
 
 	"sigs.k8s.io/kubebuilder/v3/pkg/machinery"
 
-	"github.com/operator-framework/java-operator-plugins/pkg/quarkus/v1alpha/scaffolds/internal/templates/util"
+	"github.com/operator-framework/java-operator-plugins/pkg/quarkus/v1beta/scaffolds/internal/templates/util"
 )
 
 var _ machinery.Template = &Controller{}
