+++ pkg/quarkus/v1beta/scaffolds/internal/templates/doc.go
