+++ pkg/quarkus/v1beta/scaffolds/internal/templates/gitignore.go
