+++ pkg/quarkus/v1beta/scaffolds/internal/templates/operatorfile_test.go
