+++ pkg/quarkus/v1beta/scaffolds/internal/templates/pomxml.go
