+++ pkg/quarkus/v1beta/scaffolds/internal/templates/templates_suite_test.go
