+++ pkg/quarkus/v1beta/scaffolds/internal/templates/util/file.go
