+++ pkg/quarkus/v1beta/scaffolds/internal/templates/util/file_test.go
