+++ pkg/quarkus/v1beta/scaffolds/internal/templates/util/util_suite_test.go
