+++ pkg/quarkus/v1beta/util/util.go
